+++ frontend/controllers/SiteController.php
@@ -147,27 +147,6 @@ class SiteController extends Controller
     }
 
     /**
-     * Signs user up.
-     *
-     * @return mixed
-     */
-    public function actionSignup()
-    {
-        $cities = City::find()->select('title')->indexBy('id')->column();
-        $model = new SignupForm();
-        if ($model->load(Yii::$app->request->post()) && $model->signup()) {
-
-            Yii::$app->session->setFlash('success', 'Thank you for registration. Please check your inbox for verification email.');
-            return $this->goHome();
-        }
-
-        return $this->render('signup', [
-            'model' => $model,
-            'cities' => $cities,
-        ]);
-    }
-
-    /**
      * Requests password reset.
      *
      * @return mixed
