+++ frontend/controllers/UsersController.php
@@ -3,6 +3,8 @@
 namespace frontend\controllers;
 
 use frontend\models\Category;
+use frontend\models\City;
+use frontend\models\SignupForm;
 use frontend\models\User;
 use frontend\models\UsersSearch;
 use yii\helpers\ArrayHelper;
@@ -38,6 +40,27 @@ class UsersController extends \yii\web\Controller
         return $this->render('view', [
             'model' => $model,
             'age' => $age
+        ]);
+    }
+
+    /**
+     * Signs user up.
+     *
+     * @return mixed
+     */
+    public function actionSignup()
+    {
+        $cities = City::find()->select('title')->indexBy('id')->column();
+        $model = new SignupForm();
+        if ($model->load(Yii::$app->request->post()) && $model->signup()) {
+
+            Yii::$app->session->setFlash('success', 'Thank you for registration. Please check your inbox for verification email.');
+            return $this->goHome();
+        }
+
+        return $this->render('signup', [
+            'model' => $model,
+            'cities' => $cities,
         ]);
     }
 }
